backend/requirements.txt
# --- Python Packages for FeedbackAI Backend ---
# This file is in pip format (package==version) for universal compatibility.

blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
importlib-metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
Werkzeug==3.1.3
zipp==3.23.0
Flask-Cors==6.0.1
sentence-transformers==5.1.0
chromadb==1.0.20
groq==0.31.1
Pillow==11.3.0
PyMuPDF==1.26.4 
python-docx==1.2.0
pytesseract==0.3.13